backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
asyncpg
celery
redis
pinecone
groq
google-genai
openai
anthropic
python-dotenv
